+++ python/src/transcoder.rs
@@ -249,7 +249,12 @@ impl<const OUTPUT_ENC: u8> Inner<OUTPUT_ENC> {
         )
         .map_err(to_val_err)?;
 
-        let mut encoder = CsvEncoder::new(&mut self.output, self.use_pretty_px, self.use_pretty_ts);
+        let mut encoder = CsvEncoder::builder(&mut self.output)
+            .use_pretty_px(self.use_pretty_px)
+            .use_pretty_ts(self.use_pretty_ts)
+            .write_header(false)
+            .build()
+            .map_err(to_val_err)?;
         loop {
             match decoder.decode_record_ref() {
                 Ok(Some(rec)) => {
@@ -298,12 +303,10 @@ impl<const OUTPUT_ENC: u8> Inner<OUTPUT_ENC> {
         )
         .map_err(to_val_err)?;
 
-        let mut encoder = JsonEncoder::new(
-            &mut self.output,
-            false,
-            self.use_pretty_px,
-            self.use_pretty_ts,
-        );
+        let mut encoder = JsonEncoder::builder(&mut self.output)
+            .use_pretty_px(self.use_pretty_px)
+            .use_pretty_ts(self.use_pretty_ts)
+            .build();
         loop {
             match decoder.decode_record_ref() {
                 Ok(Some(rec)) => {