+++ rust/dbn-cli/src/encode.rs
@@ -4,20 +4,36 @@ use dbn::{
     decode::{DbnMetadata, DecodeRecordRef},
     encode::{
         json, DbnEncodable, DbnRecordEncoder, DynEncoder, DynWriter, EncodeDbn, EncodeRecordRef,
+        EncodeRecordTextExt,
     },
-    rtype_dispatch, Compression, Encoding, MetadataBuilder, SType,
+    rtype_dispatch, Compression, Encoding, MetadataBuilder, SType, SymbolIndex,
 };
 
 use crate::{infer_encoding_and_compression, output_from_args, Args};
 
-pub fn encode_from_dbn<D>(decoder: D, args: &Args) -> anyhow::Result<()>
+pub fn silence_broken_pipe(err: anyhow::Error) -> anyhow::Result<()> {
+    // Handle broken pipe as a non-error.
+    if let Some(err) = err.downcast_ref::<dbn::Error>() {
+        if matches!(err, dbn::Error::Io { source, .. } if source.kind() == std::io::ErrorKind::BrokenPipe)
+        {
+            return Ok(());
+        }
+    }
+    Err(err)
+}
+
+pub fn encode_from_dbn<D>(mut decoder: D, args: &Args) -> anyhow::Result<()>
 where
     D: DecodeRecordRef + DbnMetadata,
 {
     let writer = output_from_args(args)?;
     let (encoding, compression) = infer_encoding_and_compression(args)?;
-    let encode_res = if args.should_output_metadata {
-        assert!(args.json);
+    Ok(if args.should_output_metadata {
+        if encoding != Encoding::Json {
+            return Err(anyhow::format_err!(
+                "Metadata flag is only valid with JSON encoding"
+            ));
+        }
         json::Encoder::new(
             writer,
             args.should_pretty_print,
@@ -28,24 +44,25 @@ where
     } else if args.fragment {
         encode_fragment(decoder, writer, compression)
     } else {
-        DynEncoder::new(
-            writer,
-            encoding,
-            compression,
-            decoder.metadata(),
-            args.should_pretty_print,
-            args.should_pretty_print,
-            args.should_pretty_print,
-        )?
-        .encode_decoded(decoder)
-    };
-    match encode_res {
-        // Handle broken pipe as a non-error.
-        Err(dbn::Error::Io { source, .. }) if source.kind() == std::io::ErrorKind::BrokenPipe => {
+        let mut encoder = DynEncoder::builder(writer, encoding, compression, decoder.metadata())
+            .all_pretty(args.should_pretty_print)
+            .with_symbol(args.map_symbols)
+            .build()?;
+        if args.map_symbols {
+            let symbol_map = decoder.metadata().symbol_map()?;
+            let ts_out = decoder.metadata().ts_out;
+            while let Some(rec) = decoder.decode_record_ref()? {
+                let sym = symbol_map.get_for_rec(&rec).map(String::as_str);
+                // Safety: ts_out is accurate because we get it from the metadata
+                unsafe {
+                    encoder.encode_ref_ts_out_with_sym(rec, ts_out, sym)?;
+                }
+            }
             Ok(())
+        } else {
+            encoder.encode_decoded(decoder)
         }
-        res => Ok(res?),
-    }
+    }?)
 }
 
 pub fn encode_from_frag<D>(mut decoder: D, args: &Args) -> anyhow::Result<()>
@@ -60,7 +77,7 @@ where
     }
     assert!(!args.should_output_metadata);
 
-    let mut encoder = DynEncoder::new(
+    let mut encoder = DynEncoder::builder(
         writer,
         encoding,
         compression,
@@ -72,10 +89,11 @@ where
             .stype_in(None)
             .stype_out(SType::InstrumentId)
             .build(),
-        args.should_pretty_print,
-        args.should_pretty_print,
-        args.should_pretty_print,
-    )?;
+    )
+    // Can't write header until we know the record type
+    .write_header(false)
+    .all_pretty(args.should_pretty_print)
+    .build()?;
     let mut has_written_header = encoding != Encoding::Csv;
     fn write_header<T: DbnEncodable>(
         _record: &T,
@@ -85,24 +103,10 @@ where
     }
     while let Some(record) = decoder.decode_record_ref()? {
         if !has_written_header {
-            match rtype_dispatch!(record, write_header, &mut encoder)? {
-                Err(dbn::Error::Io { source, .. })
-                    if source.kind() == io::ErrorKind::BrokenPipe =>
-                {
-                    return Ok(())
-                }
-                res => res?,
-            }
+            rtype_dispatch!(record, write_header, &mut encoder)??;
             has_written_header = true;
         }
-        // Assume no ts_out for safety
-        match encoder.encode_record_ref(record) {
-            // Handle broken pipe as a non-error.
-            Err(dbn::Error::Io { source, .. }) if source.kind() == io::ErrorKind::BrokenPipe => {
-                return Ok(());
-            }
-            res => res?,
-        };
+        encoder.encode_record_ref(record)?;
     }
     Ok(())
 }